性能优化：减小buffer初始化大小 优化结果：每秒处理100字节消息包数量由33400.1变为35568.2

+++ server/src/server/basic/timerqueue.cpp
@@ -21,7 +21,7 @@ int TimerQueue::run()
 	while(!m_taskqueue.empty()) {
 		Timer &task = *m_taskqueue.top();
 		if(task.m_expired > now) {
-			return task.m_expired - now;
+			return (int)(task.m_expired - now);
 			break;
 		}
